se agrega metodo constructor a las entitys

+++ src/main/java/com/udea/edu/co/busquedadevuelos/backendvirtual/repositories/CiudadRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.udea.edu.co.busquedadevuelos.backendvirtual.entities.Ciudad;
 
 @Repository
-public interface CiudadRepository extends JpaRepository<Ciudad, Integer> {
+public interface CiudadRepository extends JpaRepository<Ciudad, Long> {
     
 }
